// src/types.ts
export interface MarketplaceItem {
  id: string;
  title: string;
  price: string;
  category: 'electronics' | 'property' | 'vehicles' | 'fashion' | 'others';
  imageUrl: string;
  location: string;
  sellerName: string;
  sellerAvatar: string;
  description: string;
  reviews: { reviewer: string; stars: number; comment: string }[];
  created_at: string;
}